Fix issue #8. Renamed the 'dmg' target to 'release' (the old 'release' target is now 'release-build'). Default "tag" is the trunk.
Added a 'help' target that shows available make targets.

Removed the 'snapshot' targets as this is handled by 'make release' (and it
wasn't updated to work with svn anyway).

Added a 'cli' target that builds the CLI in the local tree.

Also fix regression testing when building a release: use
BUILD_PROFILE=release when making 'check'. This way we link against the
correct SDK for tests too.



git-svn-id: http://shakespeer.googlecode.com/svn/trunk@451 3d55443d-7b4c-0410-a533-cbe148fed889